HW-3: метод createAnimals теперь возвращает Map<String, List<Animal>>, Появляются две новых сущности: интерфейс AnimalRepository, Его имплементация
