+++ HW-1/src/test/java/animal/service/AnimalsRepositoryImplTest.java
@@ -1,124 +1,24 @@
-package animal;
+package animal.service;
 
+import animal.Animal;
 import animal.pet.Cat;
 import animal.pet.Cow;
 import animal.pet.Dog;
 import animal.predator.Lion;
 import animal.predator.Shark;
 import animal.predator.Wolf;
-import animal.service.AnimalsRepositoryImpl;
-import animal.service.SearchServiceImpl;
 import custexceptions.EmptyAnimalArrayException;
-import custexceptions.InvalidAnimalBirthDateException;
-import custexceptions.InvalidAnimalException;
 import custexceptions.NullAnimalArrayException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.NullSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
 import java.time.LocalDate;
 import java.util.Map;
-import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-/**
- * Класс для тестирования
- */
-class SearchServiceImplTest {
-    SearchServiceImpl ssi = new SearchServiceImpl();
+class AnimalsRepositoryImplTest {
     AnimalsRepositoryImpl ari = new AnimalsRepositoryImpl();
-
-    /**
-     * Проверка на корректность вискосного года
-     *
-     * @param year - високосный год для проверкм
-     * @throws InvalidAnimalBirthDateException - исключение "не заполнена дата рождения"
-     */
-    @DisplayName("Високосный год")
-    @ParameterizedTest
-    @ValueSource(ints = {2000, 2004, 2008, 2012, 2016, 2020, 2024})
-    void checkTrueLeapYearAnimal(Integer year) throws InvalidAnimalBirthDateException {
-        AbstractAnimal a = new Cat();
-        a.name = "Test";
-        a.birthDate = LocalDate.of(year, 1, 1);
-
-        ByteArrayOutputStream outContent = new ByteArrayOutputStream();
-        System.setOut(new PrintStream(outContent));
-
-        ssi.checkLeapYearAnimal(a);
-
-        String expectedOutput = "Test был рожден в високосный год.\r\n";
-        assertEquals(expectedOutput, outContent.toString());
-    }
-
-    /**
-     * Проверка на корректность не вискосного года
-     *
-     * @param year - не високосный год для проверкм
-     * @throws InvalidAnimalBirthDateException - исключение "не заполнена дата рождения"
-     */
-    @DisplayName("Не високосный год")
-    @ParameterizedTest
-    @ValueSource(ints = {2001, 2002, 2003, 2005, 2006, 2007, 2009})
-    void checkFalseLeapYearAnimal(Integer year) throws InvalidAnimalBirthDateException {
-        AbstractAnimal a = new Cat();
-        a.name = "Test";
-        a.birthDate = LocalDate.of(year, 1, 1);
-
-        ByteArrayOutputStream outContent = new ByteArrayOutputStream();
-        System.setOut(new PrintStream(outContent));
-
-        ssi.checkLeapYearAnimal(a);
-
-        String expectedOutput = "Test не был рожден в високосный год.\r\n";
-        assertEquals(expectedOutput, outContent.toString());
-    }
-
-    //просто для теста аннотации MethodSource, логика необязательна
-    private static Stream<Arguments> fetchData() {
-        return Stream.of(
-                Arguments.arguments(new Cat()),
-                Arguments.arguments(new Dog()),
-                Arguments.arguments(new Cow()),
-                Arguments.arguments(new Lion()),
-                Arguments.arguments(new Shark()),
-                Arguments.arguments(new Wolf())
-        );
-    }
-
-    /**
-     * Проверка на исключение InvalidAnimalBirthDateException
-     *
-     * @param a - экзмепляр класса AbstractAnimal или его потомки
-     */
-    @DisplayName("Исключение InvalidAnimalBirthDateException")
-    @ParameterizedTest
-    @MethodSource("fetchData")
-    void checkLeapYearAnimalException(AbstractAnimal a) {
-        a.birthDate = null;
-
-        assertThrows(InvalidAnimalBirthDateException.class, () -> ssi.checkLeapYearAnimal(a));
-    }
-
-    /**
-     * Проверка на исключение InvalidAnimalException
-     */
-    @DisplayName("Исключение InvalidAnimalException")
-    @ParameterizedTest
-    @NullSource
-    void checkInvalidAnimalException(Animal val) {
-        assertThrows(InvalidAnimalException.class, () -> {
-            ssi.checkLeapYearAnimal(val);
-        });
-    }
-
     private Animal[] animals() {
         Animal[] animals = {
                 new Dog("Animal1", "Breed1", 150.0d, "Character1", LocalDate.of(2001, 1, 1)),
@@ -137,8 +37,8 @@ class SearchServiceImplTest {
 
     @DisplayName("Проверка метода findLeapYearNames на корректность")
     @Test
-    void indLeapYearNamesTest() {
-        Map<String, LocalDate> map = ari.indLeapYearNames(animals());
+    void findLeapYearNamesTest() {
+        Map<String, LocalDate> map = ari.findLeapYearNames(animals());
 
         assertEquals(3, map.size());
         assertEquals(LocalDate.of(2004, 1, 20), map.get("Lion Animal4"));
@@ -148,14 +48,14 @@ class SearchServiceImplTest {
 
     @DisplayName("Проверка метода findLeapYearNames на исключение NullAnimalArrayException, EmptyAnimalArrayException")
     @Test
-    void indLeapYearNamesExceptionTest() {
+    void findLeapYearNamesExceptionTest() {
         assertThrows(EmptyAnimalArrayException.class,
                 () -> {
-                    ari.indLeapYearNames(new Animal[]{});
+                    ari.findLeapYearNames(new Animal[]{});
                 });
         assertThrows(NullAnimalArrayException.class,
                 () -> {
-                    ari.indLeapYearNames(null);
+                    ari.findLeapYearNames(null);
                 });
     }
 
@@ -174,14 +74,14 @@ class SearchServiceImplTest {
 
     @DisplayName("Проверка метода findDuplicate на исключение NullAnimalArrayException, EmptyAnimalArrayException")
     @Test
-    void indLeapYearNamesNullArrayTest() {
+    void findDuplicateExceptionTest() {
         assertThrows(EmptyAnimalArrayException.class,
                 () -> {
-                    ari.indLeapYearNames(new Animal[]{});
+                    ari.findLeapYearNames(new Animal[]{});
                 });
         assertThrows(NullAnimalArrayException.class,
                 () -> {
-                    ari.indLeapYearNames(null);
+                    ari.findLeapYearNames(null);
                 });
     }
 
@@ -235,11 +135,11 @@ class SearchServiceImplTest {
     void findOlderThanExceptionTest() {
         assertThrows(EmptyAnimalArrayException.class,
                 () -> {
-                    ari.indLeapYearNames(new Animal[]{});
+                    ari.findOlderAnimal(new Animal[]{}, 0);
                 });
         assertThrows(NullAnimalArrayException.class,
                 () -> {
-                    ari.indLeapYearNames(null);
+                    ari.findOlderAnimal(null, 0);
                 });
     }
 }
